feat(AuthModal): Auto-close modal on successful login
Add a new effect to the AuthModal component that listens for changes in the `user` prop and `isOpen` state. If the user becomes logged in (`user` is truthy) while the modal is still open (`isOpen` is true), the modal will automatically close. This improves the user experience by removing the need for manual closing after a successful login initiated from the modal.